Move IntoIterator implementation to iter.rs
Since it is an iterator trait it fits better there
